Code readability & docs
