+++ .scalafmt.conf
@@ -1,4 +1,4 @@
-version = 2.0.0-RC7
+version = 1.5.1
 
 docstrings = JavaDoc
