+++ kademlia/README.md
@@ -1,0 +1,4 @@
+# Fluence Kademlia
+
+Implementation of S/Kademlia, with a few additions.
+
